firmware: move legacy APIs to their own folder
This keeps things less crowded as we port things over. :)
