+++ firmware/greatfet_usb/legacy_apis/usb_api_DS18B20.c
@@ -5,7 +5,7 @@
 #include "usb_api_DS18B20.h"
 #include <drivers/usb/lpc43xx/usb.h>
 #include <drivers/usb/lpc43xx/usb_queue.h>
-#include "usb_endpoint.h"
+#include "../usb_endpoint.h"
 
 #include <stddef.h>
 #include <greatfet_core.h>
