+++ firmware/greatfet_usb/legacy_apis/usb_api_DS18B20.h
