+++ firmware/greatfet_usb/legacy_apis/usb_api_adc.c
@@ -5,8 +5,8 @@
 #include "usb_api_adc.h"
 #include <drivers/usb/lpc43xx/usb.h>
 #include <drivers/usb/lpc43xx/usb_queue.h>
-#include "usb_endpoint.h"
-#include "usb_bulk_buffer.h"
+#include "../usb_endpoint.h"
+#include "../usb_bulk_buffer.h"
 
 #include <stddef.h>
 #include <greatfet_core.h>
