+++ firmware/greatfet_usb/legacy_apis/usb_api_greatdancer.c
@@ -15,9 +15,9 @@
 
 #include <drivers/usb/lpc43xx/usb.h>
 #include <drivers/usb/lpc43xx/usb_standard_request.h>
-#include "usb_descriptor.h"
-#include "usb_device.h"
-#include "usb_endpoint.h"
+#include "../usb_descriptor.h"
+#include "../usb_device.h"
+#include "../usb_endpoint.h"
 #include <drivers/usb/lpc43xx/usb_request.h>
 
 #include "glitchkit.h"
