+++ firmware/greatfet_usb/legacy_apis/usb_api_logic_analyzer.c
@@ -5,13 +5,13 @@
 #include "usb_api_logic_analyzer.h"
 
 #include "sgpio.h"
-#include "sgpio_isr.h"
+#include "../sgpio_isr.h"
 
 #include <drivers/usb/lpc43xx/usb.h>
 #include <drivers/usb/lpc43xx/usb_queue.h>
 
-#include "usb_bulk_buffer.h"
-#include "usb_endpoint.h"
+#include "../usb_bulk_buffer.h"
+#include "../usb_endpoint.h"
 
 #include <greatfet_core.h>
