+++ firmware/greatfet_usb/legacy_apis/usb_api_sdir.h
